# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tarcza Polska API",
        "description": "Backend for the Tarcza Polska civil resilience platform.\n\n* `/api/v1/*` - Citizen mobile app (Flutter). Authenticate with the JWT returned by `POST /api/v1/devices`.\n* `/api/command/*` - Command Center. Authenticate with the JWT returned by `POST /api/command/login`.\n\nAll spatial payloads are GeoJSON (WGS84, `[lng, lat]` order). Every error is an `ErrorResponse`.\n`GET /map`, `GET /verifications/pending` and `GET /alerts` send an `ETag`; repeat the value in\n`If-None-Match` to get `304 Not Modified`.\n",
        "version": "1.1.0"
    },
    "paths": {
...
        "/api/v1/devices": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Register an anonymous device and obtain a JWT",
                "operationId": "post_api_device_register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterDeviceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Device created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeviceRegistered"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid body; error.violations[].field names the offending field (error.code: validation_failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/v1/devices/me": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Current device profile",
                "operationId": "get_api_device_me",
                "responses": {
                    "200": {
                        "description": "Device profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeviceProfile"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/devices/me/location": {
            "put": {
                "tags": [
                    "Devices"
                ],
                "summary": "Update the last known location of the device (no history is kept)",
                "operationId": "put_api_device_location",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLocationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location stored, H3 cell returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LocationUpdated"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "More than 30 updates a minute from this device (error.code: too_many_requests, Retry-After header)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid body; error.violations[].field names the offending field (error.code: validation_failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/devices/me/preferences": {
            "put": {
                "tags": [
                    "Devices"
                ],
                "summary": "Notification preferences (location_refresh reminders on / off)",
                "operationId": "put_api_device_preferences",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePreferencesRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Stored"
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid body; error.violations[].field names the offending field (error.code: validation_failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/devices/me/push-token": {
            "put": {
                "tags": [
                    "Devices"
                ],
                "summary": "Register / rotate the FCM push token",
                "operationId": "put_api_device_push_token",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePushTokenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Stored"
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid body; error.violations[].field names the offending field (error.code: validation_failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorViolation": {
                "required": [
                    "field",
                    "message"
                ],
                "properties": {
                    "field": {
                        "description": "Name of the body field (maps 1:1 onto the form field)",
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ErrorDetail": {
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "description": "Stable machine-readable code: bad_request, unauthorized, token_expired, forbidden, not_found, conflict, gone, verification_already_answered, verification_expired, validation_failed, too_many_requests, internal_error.\n",
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "retryAfter": {
                        "description": "Seconds until the next attempt may succeed (429 only, also sent as the Retry-After header)",
                        "type": "integer"
                    },
                    "violations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorViolation"
                        }
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorDetail"
                    }
                },
                "type": "object"
            },
            "GeoJsonGeometry": {
                "description": "GeoJSON geometry in WGS84, coordinates ordered [lng, lat].",
                "required": [
                    "type",
                    "coordinates"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "Point",
                            "Polygon",
                            "MultiPolygon"
                        ]
                    },
                    "coordinates": {
                        "description": "Point: [lng, lat]; Polygon: [[[lng, lat], ...]]; MultiPolygon: [[[[lng, lat], ...]]]",
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
...
            "DeviceRegistered": {
                "required": [
                    "deviceId",
                    "token"
                ],
                "properties": {
                    "deviceId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "token": {
                        "description": "Bearer token for /api/v1/* (valid 30 days)",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "DevicePreferences": {
                "required": [
                    "locationRefresh"
                ],
                "properties": {
                    "locationRefresh": {
                        "description": "Receive the location_refresh reminder push (default true)",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "DeviceProfile": {
                "required": [
                    "deviceId",
                    "hasPushToken",
                    "preferences"
                ],
                "properties": {
                    "deviceId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "platform": {
                        "description": "ios | android | web | simulator",
                        "type": "string",
                        "nullable": true
                    },
                    "hasPushToken": {
                        "type": "boolean"
                    },
                    "lastLocation": {
                        "description": "Point or null until the first location update",
                        "nullable": true,
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/GeoJsonGeometry"
                            }
                        ]
                    },
                    "h3Cell": {
                        "type": "string",
                        "nullable": true
                    },
                    "locationUpdatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "locationSource": {
                        "nullable": true,
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/LocationSource"
                            }
                        ]
                    },
                    "preferences": {
                        "$ref": "#/components/schemas/DevicePreferences"
                    }
                },
                "type": "object"
            },
            "LocationUpdated": {
                "required": [
                    "h3Cell"
                ],
                "properties": {
                    "h3Cell": {
                        "description": "H3 cell (resolution 9) of the stored position",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "RegisterDeviceRequest": {
                "properties": {
                    "platform": {
                        "type": "string",
                        "nullable": true,
                        "default": null,
                        "enum": [
                            "ios",
                            "android",
                            "web",
                            "simulator"
                        ]
                    },
                    "appVersion": {
                        "type": "string",
                        "nullable": true,
                        "default": null,
                        "maxLength": 32
                    },
                    "pushToken": {
                        "type": "string",
                        "nullable": true,
                        "default": null,
                        "maxLength": 4096
                    }
                },
                "type": "object"
            },
            "LocationSource": {
                "type": "string",
                "enum": [
                    "home",
                    "gps",
                    "background"
                ]
            },
            "UpdateLocationRequest": {
                "required": [
                    "lat",
                    "lng"
                ],
                "properties": {
                    "lat": {
                        "type": "number",
                        "format": "float",
                        "maximum": 90,
                        "minimum": -90
                    },
                    "lng": {
                        "type": "number",
                        "format": "float",
                        "maximum": 180,
                        "minimum": -180
                    },
                    "accuracyMeters": {
                        "type": "number",
                        "format": "float",
                        "nullable": true,
                        "default": null,
                        "minimum": 0
                    },
                    "source": {
                        "title": "Defaults to gps when omitted.",
                        "nullable": true,
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/LocationSource"
                            }
                        ],
                        "default": null
                    }
                },
                "type": "object"
            },
            "UpdatePreferencesRequest": {
                "required": [
                    "locationRefresh"
                ],
                "properties": {
                    "locationRefresh": {
                        "title": "Receive the \"are you still in this area?\" reminder when the position is stale. Default on.",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "UpdatePushTokenRequest": {
                "required": [
                    "pushToken"
                ],
                "properties": {
                    "pushToken": {
                        "type": "string",
                        "maxLength": 4096
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ],
...
}